requirements_dev.txt
-e .
-r requirements_docs.txt
-r requirements_test.txt
-r examples/requirements.txt

coverage
Django
isort
Pillow
SQLAlchemy
mongoengine
wheel
tox
zest.releaser[recommended]
freezegun
